MLOps/Image_Classification_Pipeline_Using_Vertex AI/requirements.txt
google-cloud-aiplatform
kfp
google-cloud-pipeline-components==1.0.40
google-cloud-storage
